[AIRFLOW-2715] Use region setting when launching Dataflow templates (#4139)
To launch an instance of a Dataflow template in the configured region,
the API service.projects().locations().teplates() instead of
service.projects().templates() has to be used. Otherwise, all jobs will
always be started in us-central1.

In case there is no region configured, the default region `us-central1`
will get picked up.

To make it even worse, the polling for the job status already honors the
region parameter and will search for the job in the wrong region in the
current implementation. Because the job's status is not found, the
corresponding Airflow task will hang.

Change-Id: I60de2423d0d7fdf38527b780c468eb716a6f64de